Newest files for sprint meeting on tuesday
Worked on these over the weekend. Added functionality to the GameWebsite.html file to actually open a unity game. 
Split the python server into a webserverGame.py file, which serves only the GameWebsite.html when you run it in the terminal (im running it in vscode using ngrok), and the webserverChat.py file, which serves the entire directory.

 This will allow us to showcase the chatroomTest.html ( a working chatroom to show that we can connect users together). We can also open the GameWebsite.html using webserverChat.py but it has some issues displaying connected users in the lobby, for the purpose of showcase though I think it's fine. Feel free to dig into these files and look at any mistakes I might have made as there are a few bugs that I didn't have time to iron out.