Current working HTML, Python Server Files, and Unity
All working files for our Jackbox game, includes an updated simple_client.html file, a server.py file, and the CSharp script asset to be placed into a unity assets folder. As long as you have the csharp script, you should be able to Create the buttons and input boxes and then drag them over to the GameObject containing the script. Adding the script rather than the entire Unity build, because really all you need is the CSharp script to handle inputs and communications in Unity.